Remove google3-specific test data discovery.
Use `run_test_in_package_directory` instead.

PiperOrigin-RevId: 328540950
